feat: Add conditional insertion of new books in seedDatabase()
This commit modifies the seedDatabase() function in the tools/seed.js file to conditionally insert new books into the database. It first checks if any books with the same title already exist in the database. If there are new books to insert, it adds them using the Book.create() method. Otherwise, it logs a message indicating that there is no new data to insert.

This resolves a crash bug.